+++ rules/base/stylistic.js
@@ -1,76 +1,32 @@
 module.exports = {
   rules: {
     'camelcase': ['error', { properties: 'never', ignoreDestructuring: false }],
-    'capitalized-comments': [
-      'off',
-      'never',
-      {
-        line: {
-          ignorePattern: '.*',
-          ignoreInlineComments: true,
-          ignoreConsecutiveComments: true
-        },
-        block: {
-          ignorePattern: '.*',
-          ignoreInlineComments: true,
-          ignoreConsecutiveComments: true
-        }
-      }
-    ],
+    'capitalized-comments': 'off',
     'consistent-this': 'off',
-    'func-name-matching': [
-      'off',
-      'always',
-      {
-        includeCommonJSModuleExports: false,
-        considerPropertyDescriptor: true
-      }
-    ],
+    'func-name-matching': 'off',
     'func-names': 'off',
-    'func-style': ['off', 'expression'],
+    'func-style': 'off',
     'id-denylist': 'off',
     'id-length': 'off',
     'id-match': 'off',
-    'line-comment-position': [
-      'off',
-      {
-        position: 'above',
-        ignorePattern: '',
-        applyDefaultPatterns: true
-      }
-    ],
+    'line-comment-position': 'off',
     'lines-between-class-members': ['error', 'always', { exceptAfterSingleLine: false }],
-    'max-depth': ['off', 4],
-    'max-lines': [
-      'off',
-      {
-        max: 300,
-        skipBlankLines: true,
-        skipComments: true
-      }
-    ],
-    'max-lines-per-function': [
-      'off',
-      {
-        max: 50,
-        skipBlankLines: true,
-        skipComments: true,
-        IIFEs: true
-      }
-    ],
+    'max-depth': 'off',
+    'max-lines': 'off',
+    'max-lines-per-function': 'off',
     'max-nested-callbacks': 'off',
-    'max-params': ['off', 3],
-    'max-statements': ['off', 10],
-    'max-statements-per-line': ['off', { max: 1 }],
-    'multiline-comment-style': ['off', 'starred-block'],
+    'max-params': 'off',
+    'max-statements': 'off',
+    'max-statements-per-line': 'off',
+    'multiline-comment-style': 'off',
     'new-cap': [
       'error',
       {
         newIsCap: true,
         newIsCapExceptions: [],
         capIsNew: false,
-        capIsNewExceptions: ['Immutable.Map', 'Immutable.Set', 'Immutable.List']
-      }
+        capIsNewExceptions: ['Immutable.Map', 'Immutable.Set', 'Immutable.List'],
+      },
     ],
     'no-array-constructor': 'error',
     'no-bitwise': 'error',
@@ -87,23 +43,23 @@ module.exports = {
       {
         selector: 'ForInStatement',
         message:
-          'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.'
+          'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.',
       },
       {
         selector: 'ForOfStatement',
         message:
-          'iterators/generators require regenerator-runtime, which is too heavyweight for this guide to allow them. Separately, loops should be avoided in favor of array iterations.'
+          'iterators/generators require regenerator-runtime, which is too heavyweight for this guide to allow them. Separately, loops should be avoided in favor of array iterations.',
       },
       {
         selector: 'LabeledStatement',
         message:
-          'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.'
+          'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.',
       },
       {
         selector: 'WithStatement',
         message:
-          '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.'
-      }
+          '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.',
+      },
     ],
     'no-ternary': 'off',
     'no-underscore-dangle': 'off',
@@ -114,11 +70,11 @@ module.exports = {
       'error',
       { blankLine: 'always', prev: '*', next: 'return' },
       { blankLine: 'always', prev: ['const', 'let', 'var'], next: '*' },
-      { blankLine: 'any', prev: ['const', 'let', 'var'], next: ['const', 'let', 'var'] }
+      { blankLine: 'any', prev: ['const', 'let', 'var'], next: ['const', 'let', 'var'] },
     ],
     'prefer-exponentiation-operator': 'off',
     'prefer-object-spread': 'error',
-    'sort-keys': ['off', 'asc', { caseSensitive: false, natural: true }],
+    'sort-keys': 'off',
     'sort-vars': 'off',
     'spaced-comment': [
       'error',
@@ -126,14 +82,14 @@ module.exports = {
       {
         line: {
           exceptions: ['-', '+'],
-          markers: ['=', '!'] // space here to support sprockets directives
+          markers: ['=', '!'], // space here to support sprockets directives
         },
         block: {
           exceptions: ['-', '+'],
           markers: ['=', '!', ':', '::'], // space here to support sprockets directives and flow comment types
-          balanced: true
-        }
-      }
-    ]
-  }
+          balanced: true,
+        },
+      },
+    ],
+  },
 };